backend/requirements.txt
streamlit
folium
streamlit-folium
shapely
numpy
pandas
geopandas
sqlalchemy
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
passlib[bcrypt]
python-dotenv